harvester/requirements.txt
appnope==0.1.0
attrdict==2.0.0
attrs==17.3.0
backports-abc==0.5
backports.csv==1.0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
biocommons.seqrepo==0.3.5
biopython==1.69
bioutils==0.3.3
bleach==2.1.2
bokeh==0.12.5
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
cycler==0.10.0
decorator==4.1.2
elasticsearch==5.2.0
elasticsearch-dsl==5.3.0
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
hgvs==1.1.0.post1
holoviews==1.8.4
html5lib==1.0.1
idna==2.6
inflection==0.3.1
iniherit==0.3.9
ipdb==0.11
ipykernel==4.7.0
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.5
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
kafka-python==1.3.3
lxml==3.4.1
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.2
numpy==1.13.3
pandas==0.19.2
pandocfilters==1.4.2
param==1.5.1
Parsley==1.3
pathlib==1.0.1
pathlib2==2.3.0
pexpect==4.3.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.5.2
py==1.5.2
pydash==4.1.0
Pygments==2.2.0
pyparsing==2.2.0
pysam==0.13
pytest==3.3.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
PyUpSet==0.1.1.post7
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
requests-cache==0.4.13
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.2.7
synapseclient==1.7.2
terminado==0.8.1
testpath==0.3.1
tornado==4.5.2
tqdm==4.19.5
traitlets==4.3.2
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
virtualenv==16.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.8
xxhash==1.3.0
yoyo-migrations==5.0.5
